requirements.txt
# Core dependencies
numpy>=1.21.0
pyyaml>=6.0
requests>=2.28.0
python-dotenv>=0.19.0
tenacity>=8.0.1  # For API retries
loguru>=0.6.0    # Better logging

# API clients
scholarly>=1.7.0  # Google Scholar API
semanticscholar>=0.5.0  # Semantic Scholar API

# Data processing
pandas>=1.3.0
scipy>=1.7.0

# Testing
pytest>=7.0.0
pytest-cov>=2.12.0

# Type checking
mypy>=0.910

# Documentation
sphinx>=4.5.0

# WebSocket dependencies
flask>=2.0.0
flask-socketio>=5.0.0
eventlet>=0.30.0
python-socketio>=5.0.0
d3==7.8.5
networkx>=2.5
scikit-learn>=0.24.0
tqdm>=4.60.0
fitz>=0.0.1.dev2
PyMuPDF>=1.18.14

# Security packages
cryptography>=39.0.0
pyjwt>=2.4.0